fix(runtime,zia,vm): runtime API naming audit, arithmetic conformance, and compiler enhancements
Runtime API standardization from gap report:
- Rename List.Flip → List.Reverse (with RT_ALIAS backward compat)
- Standardize Set/Bag/SortedSet: Merge→Union, Common→Intersect (with aliases)
- Rename TreeMap.Drop → TreeMap.Remove (with alias)
- Add Has() to Queue (circular buffer traversal) and Stack (linear scan)
- Fix i64→i1 return types on Canvas.IsMaximized/IsMinimized/IsFocused,
  OrderedMap.Remove, and DefaultMap.Remove

Arithmetic conformance:
- Add cross-layer arithmetic conformance tests (shift, sub-width, float div)
- Add BASIC golden tests for integer/float division and modulo edge cases
- Zia arithmetic conformance test suite

Zia compiler:
- AST printer and location scope utilities
- Optional narrowing test coverage
- Enhanced lowerer support for match expressions, collections, and methods

VM/tools:
- Debug script and debug utils enhancements
- Runner improvements and trap handling refinements
- CLI extensions for front-end invocation

All 1238 tests pass.

+++ src/runtime/collections/rt_defaultmap.c
@@ -275,7 +275,7 @@ int64_t rt_defaultmap_has(void *map, rt_string key)
     return 0;
 }
 
-int64_t rt_defaultmap_remove(void *map, rt_string key)
+int8_t rt_defaultmap_remove(void *map, rt_string key)
 {
     if (!map || !key)
         return 0;
